requirements.txt
Django>=4.2.0,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.0
django-cors-headers>=4.0.0
django-filter>=23.0
django-redis>=5.2.0
psycopg2-binary>=2.9.0
celery>=5.3.0
redis>=4.5.0
cryptography>=41.0.0
boto3>=1.28.0
django-storages>=1.13.0
Pillow>=10.0.0
python-decouple>=3.8
gunicorn>=21.0.0
whitenoise>=6.5.0
dj-database-url>=2.0.0
djangorestframework-gis>=1.0.0
drf-spectacular>=0.26.0
django-csp>=3.7
django-fsm>=2.8.0
pyotp>=2.8.0
qrcode>=7.4.0
viewflow>=2.2.0
django-jsonfield>=1.4.0

# Event System Dependencies
pika>=1.3.0  # RabbitMQ client
kafka-python>=2.0.0  # Kafka client
jsonschema>=4.17.0  # JSON Schema validation
jmespath>=1.0.0  # JSON filtering for event subscriptions

# WebSocket Dependencies
channels>=4.0.0  # Django Channels for WebSocket support
channels-redis>=4.1.0  # Redis backend for Channels
daphne>=4.0.0  # ASGI server for WebSocket support